SNAP-215: Added support to accelerate the DESCRIBE CLUSTER by caching the loadScheme() result.
Changed the read() function of a snapdbproxy_connection object to first try reading data instead of first blocking on the poll() function call. Otherwise we can end up with an extra 40ms wait for no reason.
Make use of eventfd() instead of signalfd(). The kill() is now in the snapdbproxy_connection class itself.
Fixed a bug in snap_child::connect_cassandra(), it uses two levels and if the second level fails (the connection) we must reset the f_cassandra pointer.
Also added a test and avoid the attach_to_session() if we are not connected to cassandra. It won't do anything good anyway.
Fixed the snap_thread::kill() because we cannot do a kill on a thread that's not running. Although I found about eventfd() and use that instead since it is much safer and really lightweight.
Noticed that the bio_client::write() function was not writing errors out. I added that, just in case, although like the read it goes in stderr and not the logs.
Fixed the snapdb default f_port, which now uses 4042.
Fixed another try/catch that had outdated comments.