Now threading works correctly.
I had to eliminate trying to handling things from cassandra's thread.
There is no way to join it, so I couldn't make sure it was actually
finished before destroying the old query object and creating a new one.
So this would cause all kinds of strange behavior when changing tables.

The only thing I'm unhappy about right now is that the row count is
subjective to how many rows you've read in, so there is no way the pull
down can reflect the number of rows in the table. That I can table for
now.