SNAP-215: Continuing with changing all the QCassandraLock with snap_lock.
Initialize the snap_lock in the snapwebsites.cpp initialization function (i.e. setup the communicator address/port).
Fixed the default port to snapcommunicator in the snap_lock.cpp implementation.
Noticed that the backend could access g_messager in a place where it is expected to possibly be null, applied a fix.