Removed qt4.h.
That header was there to front-end QtCore, with a GCC pragma to force
the header to be included as a system-header (compiler warnings
ignored). Since CMake provides this for free if one uses the "SYSTEM"
keyword as a parameter to "include_directories()", I decided to
implement this instead. Not to mention that we have ported to Qt5, so
even the name was wrong.