SNAP-215: Fixed snapinit so it is much more likely to properly detect that each service did quit on a STOP or equivalent.
Removed the eventfd() since we can just call shutdown() on the socket is enough to break a blocking read().
Tried 3 things in an attempt to fix the speed, but none worked: (1) went from non-blocking to blocking to avoid the poll() [kept], (2) tried with a straight socket instead of BIO_... functions, (3) tried a sched_yield(). None of these tricks made much of a difference.
Added support of the clearClusterDescription() flag in snapdbproxy. This now clears the cached version of the cluster description so it can be re-read later as required.
Added a --list command line option on snaplock so we can see what is currently locked (i.e. it is only in memory so we cannot just check a database or a file...)
Various cleanups.