SNAP-345: Added two times for locks: delay obtention and lock duration; both times were "mixed" before, possibly generating problems in snap_backend.
Changed the snaplock behavior. If you do not attempt the lock in the constructor, you do not have to get an exception on failure, lock() just returns false.
Now snap_lock objects have an is_locked() function that returns true only until the lock timeout time. After that, it returns false.
Added a "log_all" FileAppender to all logger properties so that way we can have a merged version of the logs. snapserver is not enough to debug locks and database events against other events. (however, these are not turned on by default.)
Fixed snap_lock because it would release the lock in a child (within the fork()'ed child!) where it should not be released!
Applied a fix to the snap_backend because the lock would not be released until too late in the game and it generated problems.
Noticed one missing call to remove_connection() in the rare event the fork_child() call fails in snap_backend.
The process_backend_uri() should never be called if g_child_connection is not zero, but I still added a test, just in case.
Added support for a process_timeout() callback in snap_tcp_blocking_client_message_connection, that way we can avoid the exception.
Fixed the get_timeout_date() function which did not first verify whether the f_lock_connection pointer was valid. If not valid, return 0.
Fixed a bug in activate_first_lock(). It may be called from various places, some of which are not going to clean up first. Now this very function does cleanup because otherwise we are very likely to miss an entry, which would end up blocked until something else happens.
Ameliorated the snaplock --list output so we the various identifiers separated and the timeout date.

SNAP-215: Tested, trying to remove the call to contexts() but so far it does not work at all without it (Argh!).

Made various messages normally printed by SNAP_LOG_TRACE() "disappear" to avoid the heavy load.
Fixed the snapdbproxy connection in order to properly support the order.timeout()--we have to reconnect for the timeout to be taken in account.
Removed the need of the signalfd() because the parent process can just do a shutdown() on the socket. This allows us to keep the socket in blocking mode.
Removed the div in front of the #ecommerce-cart CSS name because it is not required.
Added requirements for Qt5Script in the JavaScript plugin.
Made all the snapwebsites tests compile only if Catch.hpp is installed, even if they do not require it, it is simpler that way.
Fixed the message if Catch.hpp is not available: the error message was positive and negative when it was supposed to be negative and negative.
Various cleanups.