EX-95: Work in progress: refactoring of snap-manager.
I'm having to refactor snap-manager to use event-driven techniques to
access the database instead of the programmatic techniques that the
old libQtCassandra/thrift interface encouraged. This is a work in
progress as nothing compiles yet, and of course, the code is completely
untested. However, I need to have a check point into the git archive.