SNAP-215: Fixed bug in snapinit.cpp which would prevent execv() from working. There was a string copy happening which was not supposed to be there.
Fixed the sort of the services in snapinit. It would sort the shared_ptr<>().
Now --server-name can be defined as optional (for snapsignal).
The --snapdbproxy has to be optional because all systems do not run it. Yet, we still need it in many circumstances (i.e. snapfirewall). We still get an error in those situations if the default address:port was modified.
By allowing the --snapdbproxy to be optional, had to change the get_snapdbproxy_service() to return a null pointer.
Added a recovery tag to several items of the snapinit.xml file. Also removed the --debug option and set snapbackend to 5 minutes interval by default.
Cleaned up many .conf file (removed now unused parameters, added possible but not defined parameters.)
Fixed the port definition for cassandra in snapdbproxy.conf (9040 -> 9042).
In snapinit.conf, I commented out the server_name variable since by default we use the hostname so things should work without that name in the .conf.
Added missing log_all definition in the snapcgi.properties configuration file.
Fixed a bug in snapfirewall, it would not check whether the firewall table was properly created.
Removed "console" as a default for the snapcgi.properties. It does not make sense since it always runs in the background (started by Apache2).
Fixed bug in the addr.cpp where the port (sin_port and sin6_port fields) were used without ntohs().
Added support to have a pause upfront when creating a permanent message connection.
Properly added a set of "-------------" to all services in the log so it is easy to find where they started. Also we include the server name too (useful if you are using a log server).
Added a GOSSIP message for snapcommunicator neighbors which are expected to CONNECT to us but may not know about us. (i.e. we never sent them a GOSSIP and they do not have our IP in their neighbor list.)
Various clean ups.

SNAP-289: Updated a couple of tests to pass Coverity checks. They needed better try/catch for a new throw probably from libQtCassandra.