SNAP-215: Started work to get the GOSSIP to actually work.
Fixed the adding of a new remote_snap_communicator, it needed to be re-enabled if the existing remote_snap_communicator were already started.
Fixed the gossip_received() implementation, we need to remove the connection from the snap_communicator and also we want to remove the item from the f_gossip_ips map.
Enhanced the addr.cpp to have a way to determine the type of an IP address (i.e. private, localhost, public, etc.)
Fixed a bug in snap_tcp_server_client_connection::get_client_addr(), the define_address() returns a bool.
Added a function to get the client port and the address:port as one function.
Fixed the port defined in the f_remote_address, it was not properly swaped for the current host.
Fixed the f_remote_address in case of IPv6, the square brackest were missing.
Added many set_name() on connections so that way it makes it easier to debug the many connections.