SNAP-215: Got the LOCK to work with two computers.
Properly implemented the Flood algorithm so we can broadcast messages to all our snapcommunicators.
Added a timeout (allow 10 seconds), hop counter (allow 5 hops), message identifier (allow 1 handling), message destinations (forward only if not in list) to messages that get broadcast.
Added the name of the server to connections (especially useful for sending remote messages to specific servers.)
Added a server_name parameter to the CONNECT and ACCEPT messages.
Fixed the snaplock LOCKREADY message, we have to resend it any time a new service starts on any platform and when we start. Also send it at the right time (after we sent COMMANDS).
Fixed a bug in the DISCONNECT implementation, we cannot remove the connection if we are the one in charge otherwise we never ever reconnect
Fixed a bug in calling the send_status() from an UNREGISTER message, we cannot remove the name of the connection before calling the function.
Fixed a bug in send_status() that would use a service_connection instead of a base_connection to handle the message data and not test whether we'd get a valid pointer.
Fixed a couple of bugs in case f_started is true we have to immediately take care of the f_smaller_ips changes (i.e. make sure the connection is enabled and timeout properly setup).
Fixed many "for(auto foo : blah)" with "for(auto const & foo : blah)" to avoid a copy.
Fixed various names in the conf/*.properties files. Some were still using Snap when it was expected to be a more specific name.
Added an is_connected() function to the permanent connection implementation since we cannot really otherwise know the current status of the connection.
Increased the shutdown timeout to 10 seconds. However, we still are getting a SIGTERM when we were dealing with a remote connection.
Properly stop gossinp connections on a STOP or QUITTING event.
Various clean ups.