Websites work properly, and corrected domain delete bug.
Hooked up the websites tab properly. If you click on a domain, this tab
is correctly populated with the websites associated with that tab. Now
use the fiter feature to make sure that only websites for the selected
domain are included in the key array.

The delete bug was being too aggressive, and failing to match the
current selected domain with just the websites to delete for that
domain. Basically, it was wiping out the entire index in the websites
table. It now properly removes *only* that domain and websites
associated with it.