SNAP-215: Further modifications to correctly support stopping/restarting servers.
Strengthen the "my_address" parameter by parsing it using the addr class in the snap_communicator_server using the port of the "listen" parameter.
Applied fixes to various messages or throws so they won't happen in some normal circumstances.
Fixed bug in reading the neighbors.txt file back in. The existing code was only reading ONE line!
Fixed bug in process_connection_failed() which had to call the default implementation also (to get the set_enabled() called.)
Fixed snaplock to use DISCONNECTED instead of DIED since DIED is a local service which LOCKREADY does not apply to.
Removed the getpid() from the LOCKREADY because it adds no value, the server name has to be unique anyway.
Made sure that all GOSSIP, CONNECT, ACCEPT messages would have the IP and port information (the port is very important!)
Added code to send a DISCONNECTED message to all local services interrested whenever a process_hup() happens on a remote connection.
Moved the test on IP addresses, to know whether it is part of a computer interfaces, to the addr.cpp file.