Fixed tool build for docs.
If you create a tool in your local cmake project and want to use it
directly, never use relative path names. CMake creates a target for each
executable it creates, and the nice thing about
add_custom_command/add_custom_target, the COMMAND can be the same name
as the target and the system calculates the internal path under the hood
so you don't have to.

For example, the project name of the csspp tool is "csspp_tool". So all
that has to be done is to pass in the *project name* as the COMMAND
parameter. The rest is sorted by CMake.