Implemented new and save buttons, and added apply/revert.
Hooked in the basic functionality into snapmanager to create a new entry
in the sites table. This is a new feature and for some reason is still
buggy. However, you can alter the contents of the table farily easily. I
added an Apply and Revert button for that.

I still need to implement the remove button.

Also, now the CMake build scripts defaults to "CodeBlocks" for the CMake
generator. This way, it is automatically compatible with qtcreator.
There is a cache variable that let's you change this.