SNAP-356: Applied a few fixes to various setup files.
Fixed the log filenames to use ".log" and not "_log" in conf/snap-apache2.conf
Fixed the path to the cgi-bin, it is /usr and not /var.
Added support for snapmanager to our conf/snap-apache2.conf so it can get started.
Changed an email address with example.com instead of m2osw.com
Made apache2 a dependency of snap.cgi (it is not likely to work with any other HTTP server.)
Moved the snap-apache2.conf from snapserver.install to snapcgi.install.
Added script to install the cgi and rewrite modules in snapcgi.postinst. Also added the secure/snapcgi.log file setup.
Fixed sample of checkpermissions with more current options.
Added a warning in case the address for snapcommunicator is not 127.0.0.1 (because it won't work!)