Now adds full versioning for all Snap! dependencies.
The old perl code was pretty unsophisicated--it went through the control
file and found all occurrences of "packagename (>= <version>)". If you
failed to list a build dependency with a version, it would no longer
work. I added a bit that tests for the "Build-Depends:" stanza only and
does not modify anything outside that block. Now you can list a
dependency sans-version, and this script will put one in.