Added high/low watermark specification, --tables option and ignore tables.
When restoring a table with very large records, cassandra throws
watermark warnings if you exceed it. The high watermark is set very low
by default: 65536. I added options so that you can change the default
watermarks (low and high) to taste, enough to make the warnings go away.

Also, the "--tables" option was not hooked up, so I added the code for
that. Now you can dump/restore select tables. Also, the snapTableList
static list was not initialized, so I added that as well.

If the schema already exists (keyspace is already created), snapbackup
will now not bother trying to create the tables and context. However,
you can override this behavior with a new switch
(--force-overwrite-schema). This might be useful if you accidently
dropped a table and want it restored.