SNAP-354: Install the configuration files in snapbase.
Fixed the CMakeLists.txt of the snapcommunicator to only install the snapcommunicator configuration file (it was not included hence it never worked like that here.)
Also the snapmanagercgi would not include the CMakeLists.txt from the conf sub-directory.
Broken up the snapserver.logrotate files in each corresponding project.
Added a detectadblocker logo file.