SNAP-354: Updated dependencies, recommendations, suggestions... and made sure to add the conf/* files to packages as expected.
Added a <name>-doc package to all the project that had only one .install file (because dh_install has a bug and cannot handle a single file!)
Accordingly, added doxygen and graphviz as dependencies.
Removed the --std=c++11 from libsnapwebsites, there is such a definition in the SnapCMakeModules and we want to have --std=c++14 now.
Tried a [[deprecated]] attribute, but with --Werror, it generate errors instead of warning (duh!)
Fixed the installation of the snapfirewall XML table definition. It was not being installed at the right location.
Fixed the snaplayout dependency on snapdbproxy and not snapdb.
Fixed the iplock build dependencies, the graphviz was missing.
Fixed the Suggests in iplock ('s' was missing).