SNAP-354: Fixed the snapserver, I had the documentation, but did not add the add_subdirectory().
Fixed the few <name>.logrotate post scripts that may not have access to snapsignal (because it is part of snapcommunicator and snapbase, snapinit do not force that package to be installed.)
Added a snapwebsites.d directory. Once that shows up, I will look into implementing the functionality.
Started work on having a snapmanager daemon.