SNAP-110: Started enhancing the snapmanagerdaemon so it would understand some instructions.
Fixed the "spelling" of messenger in one message of the snaplock daemon.
Fixed the name of a variable which did not reflect what it was being used for very well.
Fixed the rules for snapmanagerdaemon. We want that binary to have the 's' flag set for user (u+s).
Put the /usr/bin/* back in since now snapmanager daemon needs to be installed.