SNAP-110: Continued to work on the snapmanager.cgi implementation.
Added support to read configuration files in a sub-directory named <path>/<project>.d/<basename> if it exists to advgetopt and libsnapwebsites.
Fixed the setgid() failure message to say 'g' and not 'u'.
Removed references to libprocps3-dev.
Removed the "server_types" from the snapcommunicator.conf and binary handling since it is really not required. It will be handled by snapmanagerdaemon instead.
Replaced the connection type by a simple enumeration: down, locale, or remote.
Fixed the set_connection_type() called in the CONNECT implementation of snapcommunicator. It was done whether we would send REFUSE or ACCEPT.
Applied a fix to several string that were not properly converted to QString as UTF-8 srings.
Added configuration files for snapmanagerdaemon.
Added libprocps4-dev to the snapmanagercgi project as we are including more things from the library it become necessary.
Various clean ups.