Fixed bug in cassview that was causing records to be overwritten.
If you used the arrow keys to navigate in the columns, it would
accidently write the value from the previous record into the next
record. This is because when loading the QDocument, the modified flag is
set to true. Now this flag is set to false on initial load, and only
writes when it has actually been modified.