Fixed issues with cassview.
Added a new feature to force a prompt for when you are modifying the
database. Added a setting to turn this off or on.

While modifying hte SettingsDialog, I found a legacy connection to
QCassandra to test the host name and port. This was throwing an
unhandled exception, which caused the application to do strange things.
This has been fixed, and it attempts to connect with QCassandraSession
instead.