SNAP-274: Added support for global singletons in Snap!
Now a cron action can also come with a GLOBAL_LOCK parameter. If set, the URI is not used to lock the process. Just the action. This makes the lock cluster wide, hence global.
Added the GLOBAL_LOCK parameter to the sendmail backend.
Removed warning from the function reading the configuration file.