Improved the connection dialog, on startup as well.
There was a small bug such that when cassview started up with server
settings to a server that is down, the problem would hard throw and
exception and fail ungracefully. This also meant that you had no way of
changing the server settings except through the configuration file.

I added a nice bit to the SettingsDlg that allow one to test the
connection, and if it fails, to gracefully handle it. Now the main.cpp
file also uses this, and if the connection is bad, it will allow you to
change it and try again.