Conversion of snaplayout to QCassandraQuery is feature complete.
The new changes compile, but for the most part are untested. This
is a work in progress.

However, I did fix a problem in advgetopt where an exception was being
thrown because the default argument had a f_name that was nullptr. I
changed this to allow for an empty argument for default arguments only.
Also, if you list a name for default arguments, "--" is not prepended in
the help output.