Fixed the filtering in the QueryModel, and added a Clear button in cassview.
The filtering was not working properly because when you use
canFetchMore() and fetchMore(), they are only called by the view when it
thinks the contents (or the geometry of the widget) has changed.
Specifically, it triggers a timer that calls fetchMore(). This happens
to work when there is no filter because when canFetchMore() is called,
the first rows are in the pending queue, and therefore are ready to be
added to the main vector and displayed by the widget.

When you turn on filtering, it may take some time before matches are
found. In the meantime, the control thinks there is nothing to fetch and
sits idle. I discovered that if I resized the main window, suddenly
entries would appear.

I don't use that method any more. I discovered that when Qt5 emits a
signal on a background thread, it arrives only in the main (UI) thread
(through the event system). What this means is that I can drive the
populating of the model from the background thread, through a signal,
and I'm assured I will only be modifying the UI from the main thread.

This means that the control is populated completely with all entries
(meaning the scrollbar accurately shows the contents instead of changing
size when you scroll to the bottom as new rows are added). It also means
that filtering works like a champ.

The downside is that it's a little slow. I had to use a 1sec timer to
slowdown the popluation, otherwise the UI is locked while the load
happens in the background.

I also tested this with snapmanager, and it seems to work great.