SNAP-110: Got a first version of snapmanager.cgi that returns a status from the snapmanagerdaemon process.
Changed the implementation of the snapmanagerdaemon concept: it will now have a thread checking the status and then sending it to other computers, constantly.
Added the snapmanagerdaemon to the snapinit.xml configuration file.
Added the daemon properties, configuration and the CGI parser to the configuration files of snapmanagercgi.
Added a logrotate file for the snapmanager.cgi and snapmanagerdaemon binaries.
Actually fixed the snapmanagerdaemon.properties file to reflect that daemon and not snapmanager.cgi.
Fixed the test of who sent the message in the snapmanagerdaemon, we have to use the get_sent_from_service() function!
Changed the snap thread FIFO with a queue instead of deque. I may regret it... but the queue is expected to be safer for a FIFO.
Added an inter-threads communication system that "exchanges" messages from the main thread and a child through a simple FIFO.
Fixed the blocking client message connection, it now exits as soon as possible if it is marked as disabled.
Applied a fix to a variable in snap_pipe_connection that was not initialized in the the class declaration.
Enhanced the XML linter invocation so it can work with .xsd files as well.
Added warnings in the CGI .properties files about the "console" breaking them and thus removal of such definitions.
Also removed the log_all definitions that were commented out.
Fixed the snap.cgi logrotate definitions, I now set "www-data" as the user/group and create the snapcgi.log files.
Replaced the parsing of the address:port in snap.cgi with the snap_addr::addr object.
Added a snapinit.xsd to verify that snapinit.xml is valid at compile time. The file will also be available to administrators in our /usr/share/snapwebsites/conf directory.
Various clean ups.