Toolbuttons instead of text-buttons, took out context drop-down, wait cursors and more.
The cassview app was made from the beginning to use the old thrift
interface into the Cassandra database. The old schema layout was fixed
for all tables: key, column1 and value. However, since we ported it to
work with cassandra-cpp-driver, which uses CQL, the layout of tables is
no longer fixed. Hence, it doesn't make any sense to try to pigeon-hole
CQL tables into this gui-layout. It only makes sense for Snap!
keyspaces.

So, I have removed the keyspace combobox from the main UI, and I have
added a new field to the SettingsDialog which allows you to type in the
name of your snap context. Presumably there will ever only be one,
unless we change that. But it's a fixed thing--something you are not
likely to change at all.

I've made some other enhancements such as replacing the QPushButtons on
the main UI with QToolButtons and adding some nice free, GTK icons.
This saves on screen real-estate.

For the models that use dbutils, those have become pointers to be
initialized each time you call doQuery(). This should save a bit of time
per row rendering.

Also, fixed the issue in QueryModel where the signal "queryFinished()"
was being sent to soon.