SNAP-110: First version showing the status of all the computers in the cluster. This is a rather fake status right now, but it gets the round trip for everything...
Properly add the status connection.
Properly start the status thread so it actually start processing.
Fixed the STOP so the status connection gets managed as expected (it would get stuck otherwise.)
Fixed the sending of the MANAGERSTATUS message so it actually does get broadcast around via snapcommunicator (the thread does not have direct access to that connection!)
Changed the behavior so we send the status only to a few select computers that run snapmanager.cgi and then we save that data to files instead of having to communicate it via a TCP/IP connection.
Added a validation of the snapmanagercgi-parse.xsl file.
Fixed the snap_thread::continue_running() function, it would return the inverse of what was expected (missing the ! operator.)