SNAp-168: Fixed the process_message() in snapcommunicator, if the server is defined and it is not us, do not interpret the message ourselves, instead forward it to that other server.
After an explicit DISCONNECT, make the system wait 15 minutes before attempting to reconnect.
Added a test to prevent broadcast a message with "*" or "?" and the use of a specific server name, since it does not make sense.
SNAP-354: Added a couple of "missing" snapbase since I moved the simplified-xslt20.dtd file.