SNAP-364: Fixed editing bugs, row loading bugs, and filtering issues.
Since turning on editing features for columns, I discovered and repaired
bugs related to this, particularly when changing a column key (if
the dbutils class throws, I needed to catch it and use a plain string).

When switching from tables, the row list wouldn't properly purge out the
old contents. For example, when switching from files back to content. Now
the models are all allocated on the heap, and destroyed when the table
is changed.

Now added overrides for the find-filter methods to use dbutils to filter
correctly. For example, you can enter the first few numbers of a file and
it will work correctly now.