SNAP-354: Fixed the cassview version by grabbing it from the changelog file.
Updated the version of some CMakeLists.txt files to use cmake version 3.5+.
Fixed the name and location of the libsnapwebsites documentation. Renamed the doxy file accordingly.
Moved the schema-for-xslt20.xsd file to snapbase/doc so it shows up along our simplified version. We cannot use it with xmllint, but it is always a good idea to have it.
Renamed the core plugins project to match the folder and use the word "_project" at the end.
Added the necessary to generate the plugins documentation (doxy file, CMakeLists.txt, footer.html).
Various cleanups.