Added type output to cassview and snapdb.
Now cassview shows the column type in the group title for the "Value"
edit box.

Likewise, snapdb will default to showing the type of the column at the
end of the output. You can turn this off with the new flag "--no-types".

Also, I added some formatting to the snapdb out, so when multiple cells
are displayed, the "name = value" pairs are spaced properly so the "="
sing matches up. This makes it easier to scan the output.