Fixed small bug with snapmanager.
When you already have a domain added, then add another one, then attempt
to switch to the "Websites" tab, multiple CQL errors are thrown and
eventually the app crashes. This is because the website field was not
set. Anyway, I check the text first, then do not issue that SELECT
statement if it is blank.