Small changes and a bug fix.
When specifying "--no-log" in the backend utility, it was throwing and
exception because the g_progname global was not being set (which console
logging needs).

Also, changed some comments in the permissions plugin to better document how to use snapbackend to
check permissions (we now need --servername), and took out an unused
var.

In the user module, it's best to use the new C++11 strong enums. They
self-document far better, and are easier to read.