SNAP-110: Intermediate checkin for work on snapmanager.cgi
Got first implementation of snapmanger.cgi host page showing name/value data (but that's not going to fly because we want to be able to extend that with time.)
Added runpath to snapinit.conf so we can create /run/snapwebsites early on (and also other daemons may not have enough permissions).
Create the cluster-status directory dynamically, in case the administrator wanted to use /run/snapwebsites/... which disappear on each reboot.
Fixed start script, the path to snapinit binary when using -r was wrong.
Added a README with some info to the snapmanager.cgi directory.
Extracted the status_file class in its own .cpp file and extended it to be able to read the fields from that class. Various other clean ups too.
Fixed the path to the home page, it has to be "/cgi-bin/snapmanager.cgi"