SNAP-373: Columns were "out of sync." This has been repaired.
What was happening is in the TableModel, when the data role was
"Qt::UserRole", the model would look up the index directly from the
underlying vector, instead of translating the index from the sort map
first. So what would happen is that the row we are looking at does not
show correctly in the columns.