Implemented proper sorting into snapmanager and fixed a bug.
I cut-and-pasted code from cassview which allows for proper sorting of
keys in tables. This applies to DomainModel, TableModel and
WebsitesModel.

The bug: when attempting to edit the rules for a domain, a pop-up
message box would tell you that you're trying to add a domain which is
already there, and of course that's silly--the operation would be edit
instead. I remarked off the message boxes for now.