EX-112: The anti-virus did not start properly because the std::string::c_str() would be lost. I just use strdup() now.
Noticed that the return code from clamscan says 1 for "has virus" and 2 for "has errors". We only react on 1 now.

EX-111: The content status is now only using the current state and not the "working" information. It may have worked with the USING TIMESTAMP or the timestamp generator which ensures each access has a different timestamp...
The one process that was really using the is_working() test was the cloning. I do not foresee any problems with that in finball because it is not used much.
The CREATE state is now the row without a status and the content::primary_owner field.
Removed some dependencies on controlled_vars.

SNAP-61: Queries with a range against the column can use '<' instead of '<='. We get the same effect and it looks less like a lie.
Added USING TIMESTAMP to all our INSERT calls. We may remove those and use the timestamp generator instead. TBD

EX-69: All the loops in the snap_backend.cpp file still had the clearCache() inside the for(;;) loop, which is wrong since we use CQL.

SNAP-110: Started work to have a library in snapmanagercgi. It all compiles but is not tested at all.
Added a message so the snapmanagerdaemon can query the public IP address of snapcommunicator.
Broadcast the UNREACHABLE message whenever a snapcommunicator server cannot be reached.
Enhanced the test to know whether lock messages are being sent and avoid the TRACE in snapcommunicator if so.
Enhanced the snaplock to display the obtension timeout when a lock is still waiting to be obtained.

Various cleanups.