SNAP-347: Added defered stop to wait until "children" services have stopped.
When a service needs to stop, it must also make sure any services it
requires to be running are also stopped.

Example: if A has a dependency on B and C, if we want to stop A, we must
first signal B and C to stop.

This is accomplished by testing to see if the service is dependent on
anything else first, then waiting (but leaving the timeout active) until
all "children" services have stopped. I put "children" in quotes because
they aren't truly child process--however, they are reverse dependencies,
if you will.