iSNAP-110: Added support for plugins in snapmanager.cgi/daemon so we can add as many entries as necessary even from external sources.
Added the missing stylesheet in the .conf file. I certainly had it in the old version but it did not make it in the joined version.
Removed the wait on the public IP in manager_status, instead I start the thread after I received that message.
Replaced the loop over internal (very limited!) functions to sending the retrieve_status() signal to the plugins.
Moved the existing functions to the "self" plugin which represents the one plugin that will check the snapmanager.cgi/daemon environment.
Added the SERVER_PUBLIC_IP to the list of commands we accept in snapmanagerdaemon.
Added the --debug command line option which somehow I missed duplicating.
Removed support for user to select which plugins to load, instead we always load them all.
Various cleanups.