NAP-110: Got the plugins output used as before. Next, to implement the display within the plugin itself.
Changed the status_file class with server_status and a sub-class status_t to define each field. This gives us many possibilities.
Made updates as required to support the new server_status class (replacing the server_status_t typedef.)
Broke up the log_config in two so we have one for the CGI and one for the daemon.
Added a missing return for the PUBLIC_IP and SERVER_PUBLIC_IP messages.
Separated the log_server definition so we can properly document it too.
Readded the two .properties as I had earlier.
Actually implemented support for log_server in snapmanager.cgi/daemon.