Moved switch statement messages into std::function map.
Instead of using the prematurely-optimized switch statement to look only
at the first letter, then do a string comparison (for snapinit, this
does not save us much since the traffic is low), I converted the entire
structure to using STL maps--one for UDP and one for TCP. This way, you
look up the command in the map, then call the function stored in the
map. Lamba functions make this possible, thanks to C++11!