EX-347: Fixed a bug where the snapcommunicator is never being marked as registered.
Added denpendencies to all that need them in snapinit.xml.
Cleaned up and added information about the new tags to the snapinit.xml documentation.
Removed the sending of the SAFE message from snapdbproxy and snaplock since we now have proper dependencies in snapinit.
Removed the two corresponding <safe> tags from snapinit.xml.
Extended the <meta> tag so it supports charset as an attribute and allows for dynamically defining attributes.
Fixed the charset specification so it matches the HTML5 version if used by such a browser. (the Content-Type should be ignored)
Added the version of Snap! Websites to the generator meta tag.
Added the application-name meta tag.
Added the msapplication-config meta tag so we can at least home that most will not check by setting it to none.
Changed the output of snapmanager.cgi so it is clearly HTML5.
Added the noimageindex and nosnippet to the list of supported robots.txt meta tags.
Removed references to CfDef in libQtCassandra / table.cpp