SNAP-110: Implemented the UNREACHABLE by setting the corresponding computer to "down".
Added code to count the number of errors and warnings and save that information in the header.
Added a function to read the list of computers defined on a frontend snapmanager computer.
Added a column with the total number of errors and warnings found on a computer.
Added classes for warnings, errors, computer down cases.
Added the != operator for snap_addr::addr.
Fixed the read_header() which would not save the last field it read, field that was then lost if the file was written back out.
Started changing the self status to better match what we are expected to use.