SNAP-110: Added support for the plugins to generate their output on a per field basis.
If a field does not have a specific form attached to it, the default kicks in (i.e. show the value as is.)
Added a column with the name of the plugin managing a certain field so that way it makes it crystal clear who does what.
Added files about the snap bundles. These will be used to know what can be installed (because snapmanager.cgi cannot know what's available without some form of an external list that can be dynamically created.)
Various clean ups.